Interface HtmlToPdf defined, implementation of conversion by Adobe SDK moved to descendant class HtmlToPdfByAdobeSdk.
#51
