+++ PdfService/Controllers/PdfController.cs
@@ -6,6 +6,7 @@ using System.Net;
 using System.Web.Mvc;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using EnterpriseServices.HtmlToPdf;
 
 namespace EnterpriseServices.Controllers
 {
@@ -81,7 +82,7 @@ namespace EnterpriseServices.Controllers
         /// </summary>
         /// <param name="fileName">Content file name.</param>
         /// <returns>URL.</returns>
-        private string GetContentUri(string fileName)
+        public string GetContentUri(string fileName)
         {
             UriBuilder urlBuilder = new UriBuilder(Request.Url.AbsoluteUri)
             {
@@ -96,7 +97,7 @@ namespace EnterpriseServices.Controllers
         /// </summary>
         /// <param name="fileName">File name.</param>
         /// <returns>Path.</returns>
-        private string GetContentPath(string fileName)
+        public string GetContentPath(string fileName)
         {
             return Path.Combine(Server.MapPath("~"), "Content", fileName);
         }
@@ -157,123 +158,15 @@ namespace EnterpriseServices.Controllers
             }
         }
 
-        /// <summary>
-        /// Requests external service for conversion original URL to PDF.
-        /// </summary>
-        /// <param name="url">Original URL.</param>
-        /// <returns>File name.</returns>
-        private string RequestConvertingService(string url)
-        {
-            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12 | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls | SecurityProtocolType.Ssl3;
-
-            UriBuilder uriBuilder = new UriBuilder(Request.Url.AbsoluteUri)
-            {
-                Scheme = "https",
-#if DEBUG
-                Host = "localhost",
-                Port = 44379,
-#else
-                Host = "adobesdk.azurewebsites.net",
-                Port = 443,
-#endif
-                Path = "pdf/document",
-                Query = "url=" + url
-            };
-
-            string request = uriBuilder.ToString();
-            HttpWebRequest req = WebRequest.CreateHttp(request);
-            req.Timeout = 30000;
-            req.ContentType = "application/json";
-            req.Method = "GET";
-
-            using (WebResponse res = req.GetResponse())
-            {
-                using (StreamReader streamReader = new StreamReader(res.GetResponseStream()))
-                {
-                    JObject jObject = JObject.Parse(streamReader.ReadToEnd());
-
-                    if (jObject["urlToPdf"].Type == JTokenType.Null)
-                    {
-                        if (jObject["message"].Type != JTokenType.Null)
-                        {
-                            //  Propagates an exception.
-                            throw new Exception(jObject["message"].ToString());
-                        }
-                        else 
-                        {
-                            throw new Exception("Converting service had returned empty URL to PDF.");
-                        }
-                    }
-
-                    return jObject["fileName"].ToString();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Copies data from stream to local file.
-        /// </summary>
-        /// <param name="dataStream">Data stream.</param>
-        /// <param name="filePath">Path to local file.</param>
-        private void CopyStreamToFile(Stream dataStream, string filePath)
-        {
-            if (dataStream != null)
-            {
-                using (FileStream fileStream = System.IO.File.Create(filePath))
-                {
-                    dataStream.CopyTo(fileStream);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Copies file from converting service to local folder and deletes it.
-        /// </summary>
-        /// <param name="fileName">Name of converted file.</param>
-        private void GetFileFromConvertingService(string fileName)
-        {
-            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12 | SecurityProtocolType.Tls11 | SecurityProtocolType.Tls | SecurityProtocolType.Ssl3;
-
-            UriBuilder uriBuilder = new UriBuilder(Request.Url.AbsoluteUri)
-            {
-                Scheme = "https",
-#if DEBUG
-                Host = "localhost",
-                Port = 44379,
-#else
-                Host = "adobesdk.azurewebsites.net",
-                Port = 443,
-#endif
-                Path = "pdf/file",
-                Query = "fileName=" + fileName + "&delete=true"
-            };
-
-            string request = uriBuilder.ToString();
-            HttpWebRequest req = WebRequest.CreateHttp(request);
-            req.Timeout = 30000;
-            req.ContentType = "application/pdf";
-            req.Method = "GET";
-
-            using (WebResponse res = req.GetResponse())
-            {
-                using (Stream dataStream = res.GetResponseStream())
-                {
-                    string pdfFilePath = GetContentPath(fileName);
-                    CopyStreamToFile(dataStream, pdfFilePath);
-                }
-            }
-        }
-
-        /// <summary>
+                /// <summary>
         /// Returns URL to PDF file.
         /// </summary>
         /// <param name="url">Original URL</param>
         /// <returns>URL to PDF.</returns>
         private string ConvertByAdobeSdk(string url)
         {
-            string fileName = RequestConvertingService(url);
-            GetFileFromConvertingService(fileName);
-            return GetContentUri(fileName);
+            HtmlToPdfByAdobeSdk converter = new HtmlToPdfByAdobeSdk(this);
+            return converter.GetUrlToPdf(url);
         }
 
         [AllowAnonymous]
